server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pgvector==0.3.5
pytest==8.3.3
